--- ee/query-service/.dockerignore
@@ -1,4 +1,0 @@
-.vscode
-README.md
-signoz.db
-bin
