--- pkg/query-service/.dockerignore
@@ -1,3 +1,0 @@
-.vscode
-README.md
-signoz.db
